Modify the account data structure and add a rotation relayer to submit transactions

+++ contracts/solana/relayer-hub/programs/relayer-hub/src/states/hub.rs
@@ -9,7 +9,7 @@ pub struct Config {
     /// Initialization flag, used to mark whether the global configuration has been initialized.
     pub initialized: bool,
     /// The total number of transaction pools.
-    pub tx_pool_number:u64,
+    pub tx_pool_number:u16,
 }
 
 impl Config {
